refactor: MutationCall.frequency becomes a float
It was `DecimalField(max_digits=5, decimal_places=4)`, so the column kept four decimal places
and threw the rest away. breseq's polymorphism mode reports more than that --
`frequency=8.39314286e-01` is a real value out of the fixtures -- and that call was stored as
`0.8393`.

**The precision was never lost from the row, only from the column.** The same import already
wrote the full float into `Mutation.supplemental_data`'s GenomeDiff record;
`test_scientific_notation_survives_as_a_value` asserted both halves side by side, `0.839314286`
in the JSON and `0.8393` beside it. The degraded copy was the queryable, filterable one, which
is backwards. That test is now the one that says the column keeps everything.

Nothing has to show the extra digits and nothing does: every display site already formats
explicitly -- `"%.2f"` in the mutation table, `"%.1f%%"` on the Samples page, `"%2f"` in the
interop payload -- and no template renders the value raw. Nothing aggregates or orders on it
either; the only query use is `aledb_filter`'s two bounds.

Two pieces of machinery existed only to work around `Decimal` and go with it. `history.py`
stringified the frequency into its snapshot and rebuilt it with `Decimal(...)`, because JSON
cannot carry a `Decimal`; JSON carries a float natively and round-trips it exactly.
`_call_kwargs` still coerces through `float()`, because every changeset row written before
this holds a string and restoring one would put a string in a float column.

`aledb_filter` built its cutoff bound as a `Decimal` so that "20% means exactly 0.2000" was
true rather than nearly true. That claim still has to hold and now holds for a different
reason -- both sides of the comparison are the same IEEE double, with no cross-type coercion
left to get wrong. The comment says so, and there is finally a test:
`test_a_call_at_exactly_the_floor_is_kept`, verified against `__lte`.

**Two visible consequences, both accepted rather than overlooked.**

`aledb-phylogeny` encodes `>= 1.0` as derived and anything below as ambiguous. The column used
to round 0.99997 up to 1.0000, so such a call arrived as fixed; it now arrives whole and is
ambiguous, and trees over polymorphism-mode data can differ. That is the honest reading -- a
call that is not 1 is not fixed -- and the old answer was an accident of a column width rather
than a decision. Thresholds deliberately unchanged, said so at `_encode`, and pinned.

The same call reads `100.0%` and shaded polymorphic on the Samples page where it used to read
`100%` and unshaded. Display rounds and phylogeny does not, so one page can call a mutation
fixed while the matrix calls it ambiguous. Correct in both places, and worth knowing.

Verified by importing a polymorphism-mode drop rather than only by suite: 0.839314286 and
0.99997 land in the column whole and match the JSON beside them, the Samples page shows
83.9% and the mutation table 0.84 with no float reaching either page, and a call at exactly
0.2 survives a 20% floor and is dropped by a 21% one.

Core 1693 tests green; assembled 1880.

+++ aledb_mutation_editor/tests/base.py
@@ -10,7 +10,6 @@ is what issues the ProjectAccess row `can_add_experiment_filter` consults -- the
 `aledb_sample.tests.test_table_actions` does it.
 """
 
-from decimal import Decimal
 
 from django.contrib.auth.models import User
 from django.test import TestCase
@@ -75,7 +74,7 @@ class EditorTestCase(TestCase):
             present=present,
             source="breseq",
             evidence={"wt_reads": 10, "mutated_reads": 30},
-            frequency=Decimal(frequency))
+            frequency=float(frequency))
 
     # --- assertions the suites share ------------------------------------------------------
